Added OLT adapter restart test
Adds a test to restart the open olt adapter after the
ONU is authenticated and then completes to ONU DHCP
after the adapter is restarted.

Updates the core restart test to only loop over ONUs
for ONU specific steps as opposed to restart the core
for each ONU.

Addresses a bug in the voltha.robot library to ensure
the port forward process is seen by ps.

Change-Id: Ief16c9b6127267ea58fe4f4690b2c2868ad28664