Update patch set 8
Patch Set 8:

Hi all,
I've got a lot of problems upload this patch to gerrit and there went still something wrong. I did not change libraries/onos.robot, I did not add it and I don't know, why it is uploaded?!
Furthermore the both data files where only updated but the previous version will not shown???
I updated my commit message to "WIP[VOL-2533] Use BBSim to test functionality with multiple OLTs", but it is not shown!
Working tools would be helpful!
But back to the patch itself. I copied the keywords "Setup Suite" "Setup Test" and "Perform Sanity Test" (rename it to "Do Sanity Test") from libraries/utils.robot and made my changes to prevent impacts to the single OLT testcases.
I restructured the 'hosts' part in multiple-bbsim-kind.yaml to the same structure like 'olts'. So we got a list of hosts with src and dst. That was an assumption by me. If I'm wrong please correct me. But with this way I was able to run in a loop over all onus per olt.
From my point of view it is not efficiently to use global variables in the keywords of libraries. We should use more arguments. I started in "Do Sanity Test" with the arguments 'host_src' and 'host_dst'. This is essential for the outlook to keep it common for single olts tests.
Please remind it is WIP. The testcase runs only once, then a voltha dowwn and up again is needed. Currently I don't know is it an issue of the testcase (teardown???) or a voltha!
The hint by Hardik to reformat from BBSM01000001, BBSM02000001,... to BBSM00010001, BBSM00020001,... was not needed, but I did it.
These small remarks for some clarification.
Thanks Torsten

Patch-set: 8
Label: Code-Review=0