Update patch set 9
Patch Set 9:

@HARDIK thanks for review. I correct all the indentations. My editor exchange 4 spaces to a tab. I also took over the suggested IF-ELSEIF- ELSE statements. Thanks for the hint.
The both modes should never run at onces. The intention of this test is to check the progress of the opnonu-go-adapter implementation, especially by the developers of the adapter, so they decide which mode will be tested. Both modes were required by the developers.
The common test setup of the library utils.robot can not be used due by this way we would not catch the first state. So we decide for an own test setup.

Patch-set: 9
Reviewer: Gerrit User 1000643 <1000643@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0