Update patch set 3
Patch Set 3:

Suite Teardown is commented out, as requirement of the openonu-go-adapter developer, to have the teardown as an own testcase. Therefore it is moved to testcase 'clean up and check' and was extended by the 'Pause Execution' keyword to stop the test if needed (for e.g. watch the resources,...) and a further check no onu port is enabled in onos,

Patch-set: 3