Update patch set 5
Patch Set 5:

With patch #5 a solution was found with conditional 'Import Library' so there is no need to install package 'tkinter' anymore! This is only needed for manual tests on our local machines. No action on onf sites are needed anymore.
'Suite Teardown' is back again, but in a modified way according the requirements of the openonu-go-adapter developer.
Thanks
Torsten

Patch-set: 5