Add "isDhcpRequired" is true for HSI service of TT workflow in this commit.
isDhcpRequired is set to true for VoIP and VoD flow per TT workflow requirement.
As a result ACL is installed on the PON to trap DHCP packets to host.
The DHCP packets of HSI service are also trapped to host as result
of this config. But ONOS drops the DHCP packets as it does have
the dhcp-relay app configuration to accept DHCP packets for HSI
service.
As a workaround, for ONOS to accept HSI service DHCP packets, the isDhcpRequired
is set to true for HSI service.
A long term solution will need to be worked out the agent/BAL to fix this issue.

Change-Id: I743f704bbcb03d1a7da0aa585689fe63d8823e31