Update patch set 3
Patch Set 3:

(9 comments)

The tests suite is developed for ATT workflow only. To extend it for DT workflow should be possible but not goal of the corresponding stories here. Especially the fact to reach omci-flows-pushed in DT workflow needs to add subscriber before and there are different flows to check than in ATT workflow. An own test suite (and corresponding story) seems to be the easier way than to use a workflow variable. That's my opinion/suggestion. Let's discuss.

Patch-set: 3