Update patch set 1
Patch Set 1:

(1 comment)

As I mentioned in my comment below, I would like to prefer the general usage of an already opened ssh connection instead always open a new one. But this should be a separate feature and should be put in common suite setup.
The checks themselfs look good for me.

Patch-set: 1
CC: Gerrit User 1000643 <1000643@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>