optimize keyword Wait for ONU Adapter Reconcile
After a lot of different test-scenarios to detect reconcile is finished, I decide
to use the way as it is with patchset 4:
1st: Wait until communication to openonu-adapter is established by help of lastcommunication
2nd: Wait until Reconcile is finished for all ONU devices by help of a simple voltctl request
     (Here I used onuimage list, if there is a better suggestion, don't hesitate to contact me)
3rd: Wait for expected oper-state
So this result is a mix of Matteos suggestion and a former test patch by me added by communication
establishment check. Helping comments and further suggestions are welcome.

Change-Id: Ib9e97f2e3662ef34fd658e89270e35ba44f3ee01