Add debugging output to Validate Device keyword
All output of 'voltctl device list -o json' is now logged in the default
Robot logs (what ends up in the log.html file). If the DEBUG environment
variable is set as True, is should also log it to the console (just a
test to check if this works correctly with Jenkins env vars).

These changes will hopefully shed a light on the failures in the
"build_berlin-community-pod-1-gpon_1T8GEM_voltha_DT_master_test"
pipeline.