Add more output to better debug failing tests
Add more logs to console in general.

For the Validate Device keyword, all output of
'voltctl device list -o json' is now logged in the default Robot
logs (what ends up in the log.html file). If the DEBUG environment
variable is set as True, is should also log it to the console (just a
test to check if this works correctly with Jenkins env vars). This
change will hopefully shed a light on the failures in the
"build_berlin-community-pod-1-gpon_1T8GEM_voltha_DT_master_test"
pipeline.

Change-Id: I2a861082d4110340ec3b3c2f4a32677b9b2fa661