Fix iperf command for multicast in utils library
* If the iperf command is executed as a daemon (-D), it will not
output to the file specified (>>). This commit removes the daemon
flag and runs the command in the background.
* When running the iperf process in the background with &, when the
Robot tests close the SSH connection the process is no longer running,
so a nohup has been added to the BNG and RG iperf commands.
* Retry multicast test until 300s timeout, since in some executions
it seems to work at the second try.

Signed-off-by: Cristina de Francisco <cristina.defrancisco@bisdn.de>
Change-Id: I06dfda9d865ceed2cbe0c4633b8a7920c0afc799